Replacing mass/tonnage related internal data with doubles

+++ src/megameklab/com/ui/Infantry/StatusBar.java
@@ -97,7 +97,7 @@ public class StatusBar extends ITab {
     public void refresh() {
 
         DecimalFormat roundFormat = new DecimalFormat("#.##");
-        float currentTonnage;
+        double currentTonnage;
         int bv = getInfantry().calculateBattleValue();
         int currentCost = (int)Math.round(getInfantry().getCost(false));
